requirements.txt
# ScamShield AI - Dependencies
# Version: 1.0.0

# =====================================================
# Core AI/ML Models
# =====================================================
#torch==2.1.0
#transformers==4.35.0
#sentence-transformers==2.2.2
#spacy==3.7.2
#huggingface_hub>=0.19.0  # For model downloads without transformers


transformers==4.35.0
sentence-transformers==2.2.2
spacy==3.7.2
huggingface_hub>=0.16.4,<0.18

# =====================================================
# Agentic Framework (LangChain + LangGraph)
# =====================================================
langchain==0.1.0
langchain-core>=0.1.0  # Core LangChain messages and utilities
langgraph==0.0.20
langchain-groq==0.0.1
langsmith>=0.0.77  # Compatible with langchain 0.1.0
groq>=0.4.0  # Groq API client

# =====================================================
# API Framework (FastAPI)
# =====================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# =====================================================
# Database Clients
# =====================================================
chromadb==0.4.18
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.23

# =====================================================
# NLP Utilities
# =====================================================
langdetect==1.0.9
nltk==3.8.1

# =====================================================
# Monitoring & Metrics
# =====================================================
prometheus-client==0.19.0

# =====================================================
# Utilities
# =====================================================
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3

# =====================================================
# Testing
# =====================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# =====================================================
# Development (optional)
# =====================================================
# black==23.11.0
# isort==5.12.0
# mypy==1.7.0
# flake8==6.1.0
